refactor: split jest utils into multiple files
